refactor : 설문 통계 페이지 리팩토링
