+++ src/components/stats/statsHeader/components/StatsMainHeader.jsx
@@ -1,7 +1,9 @@
 import React from "react";
-import styled from "styled-components";
-import fonts from "../../../../styles/fonts";
 import { useSelector } from "react-redux";
+import styled from "styled-components";
+
+import fonts from "../../../../styles/fonts";
+import LargeRoundDiv from "../../../common/LargeRoundDiv";
 
 const StatsMainHeader = () => {
   const statsList = useSelector((state) => state.stats.stats);
@@ -13,11 +15,13 @@ const StatsMainHeader = () => {
           <StatsSummeryContainer>
             <h1>{statsList.surveyTitle}</h1>
             <PeriodContainer>
-              <RoundDiv>
-                {statsList.status === "NOT_START" && "시작전"}
-                {statsList.status === "IN_PROCEED" && "진행중"}
-                {statsList.status === "DONE" && "마감"}
-              </RoundDiv>
+              <LargeRoundDiv
+                divText={
+                  (statsList.status === "NOT_START" && "시작전") ||
+                  (statsList.status === "IN_PROCEED" && "진행중") ||
+                  (statsList.status === "DONE" && "마감")
+                }
+              />
               <div>
                 <p>진행기간</p>
                 <h5>
@@ -34,7 +38,7 @@ const StatsMainHeader = () => {
             <h4>{statsList.totalQuestion}</h4>
           </div>
           <div>
-            <h5> 응답자 수 </h5>
+            <h5>응답자 수</h5>
             <h4>{statsList.totalParticipant} </h4>
           </div>
         </SideStatsContainer>
@@ -44,32 +48,35 @@ const StatsMainHeader = () => {
 };
 
 const Container = styled.div`
-  width: 100vw;
-  border-bottom: 1px solid #e1e1e1;
   display: flex;
   justify-content: center;
+
+  width: 100vw;
+  border-bottom: 1px solid #e1e1e1;
 `;
 
 const MainHeaderContainer = styled.div`
   display: flex;
-  padding: 0 3rem 3.4rem 3rem;
+
   width: 100%;
   min-width: 800px;
   max-width: 1200px;
+  padding: 0 3rem 3.4rem 3rem;
 `;
 
 const SummeryContainer = styled.div`
   flex: 1;
   display: flex;
   flex-direction: column;
+
   margin-top: 8.3rem;
   h4 {
+    margin: 0.8rem 0 0 0;
+
     ${fonts.Body3}
     font-weight: 500;
     font-size: 1.6rem;
     line-height: 1.9rem;
-    margin: 0;
-    margin-top: 0.8rem;
   }
 `;
 
@@ -87,7 +94,6 @@ const StatsSummeryContainer = styled.div`
 
 const PeriodContainer = styled.div`
   display: flex;
-
   div {
     display: flex;
     flex-direction: column;
@@ -109,21 +115,6 @@ const PeriodContainer = styled.div`
   }
 `;
 
-const RoundDiv = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  height: 2.6rem;
-  ${fonts.Body1}
-  font-weight: 500;
-  font-size: 1.4rem;
-  line-height: 17px;
-  background: ${({ theme }) => theme.gray6};
-  border-radius: 2.4rem;
-  padding: 0 2rem;
-`;
-
 const SideStatsContainer = styled.div`
   display: flex;
   margin-top: 11.8rem;
@@ -133,23 +124,23 @@ const SideStatsContainer = styled.div`
     justify-content: center;
     align-items: center;
 
-    border-radius: 3.95rem;
     width: 11.7rem;
     height: 6.2rem;
     background: ${({ theme }) => theme.lightMainColor};
+    border-radius: 3.95rem;
     h5 {
+      margin: 0;
       ${fonts.Body3}
       font-weight: 500;
       font-size: 1.3rem;
       line-height: 1.6rem;
-      margin: 0;
     }
     h4 {
+      margin: 0;
       ${fonts.Body1}
       font-weight: 500;
       font-size: 1.6rem;
       line-height: 1.9rem;
-      margin: 0;
     }
     &:nth-child(1) {
       margin-right: 2.2rem;
