+++ src/components/stats/statsMainList/components/Stats.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+
 import LongDescriptiveStats from "./typeOfStats/LongDescriptiveStats";
 import MultiChoiceStats from "./typeOfStats/MultiChoiceStats";
 import RankStats from "./typeOfStats/RankStats";
