+++ src/components/stats/statsMainList/components/StatsCategory.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { useSelector } from "react-redux";
 import styled from "styled-components";
+
 import CreateFormCard from "../../../createForm/createFormSide/components/CreateFormCard";
 
 const StatsCategory = () => {
