+++ src/components/stats/statsMainList/components/typeOfStats/SlideStats.jsx
@@ -1,15 +1,15 @@
 import React from "react";
 import styled from "styled-components";
-import fonts from "../../../../../styles/fonts";
 import ApexCharts from "react-apexcharts";
+
+import Title from "../Title";
+import StatsHeaderContainer from "../statsContainer/StatsHeaderContainer";
 
 const SlideStats = ({ stats }) => {
   const volume = stats.volume;
   let answerList = stats.satisfactionList?.map((answer, index) => {
     return { name: index - volume, data: [answer] };
   });
-
-  console.log(stats);
 
   let answerNameList = [];
   for (let i = 0; i < volume * 2 + 1; i++) {
@@ -90,15 +90,14 @@ const SlideStats = ({ stats }) => {
 
   return (
     <Container>
-      <Header>
-        <TitleContainer>
-          <p>{stats.questionType}</p>
-          <h2>
-            {stats.questionNum}. {stats.questionTitle}
-          </h2>
-          <h4>{stats.questionSummary}</h4>
-        </TitleContainer>
-      </Header>
+      <StatsHeaderContainer>
+        <Title
+          questionType={stats.questionType}
+          questionNum={stats.questionNum}
+          questionTitle={stats.questionTitle}
+          questionSummary={stats.questionSummary}
+        />
+      </StatsHeaderContainer>
       <ChartContainer>
         <ApexCharts
           options={options.options}
@@ -117,55 +116,25 @@ const SlideStats = ({ stats }) => {
 };
 
 const Container = styled.div`
+  position: relative;
   width: 47.2rem;
   height: 41.6em;
-  background: #ffffff;
+
+  background: ${({ theme }) => theme.backgroundColor};
   box-shadow: 0px 0px 7px 3px rgba(0, 0, 0, 0.25);
   border-radius: 10px;
-  position: relative;
 `;
 
 const ChartContainer = styled.div`
   margin-top: 8rem;
 `;
 
-const Header = styled.div`
-  display: flex;
-  justify-content: space-between;
-  padding: 3rem 3rem 0 3rem;
-`;
-
-const TitleContainer = styled.div`
-  p {
-    ${fonts.Body3}
-    font-weight: 400;
-    font-size: 1rem;
-    line-height: 1.2rem;
-    margin: 0;
-  }
-  h2 {
-    ${fonts.Body1}
-    font-weight: 700;
-    font-size: 1.6rem;
-    line-height: 1.9rem;
-    margin: 0;
-    margin-top: 0.6rem;
-  }
-  h4 {
-    ${fonts.Body3}
-    font-weight: 400;
-    font-size: 1.2rem;
-    line-height: 1.4rem;
-    margin: 0;
-    margin-top: 0.7rem;
-  }
-`;
-
 const LabelContainer = styled.div`
+  position: absolute;
   display: flex;
   justify-content: space-between;
+
   padding: 0 3rem 0 3rem;
-  position: absolute;
   width: 100%;
   top: 25.5rem;
 `;
