+++ src/components/stats/statsMainList/components/typeOfStats/StarStats.jsx
@@ -1,7 +1,10 @@
 import React from "react";
 import styled from "styled-components";
-import fonts from "../../../../../styles/fonts";
 import ApexCharts from "react-apexcharts";
+
+import fonts from "../../../../../styles/fonts";
+import Title from "../Title";
+import StatsHeaderContainer from "../statsContainer/StatsHeaderContainer";
 
 const StarStats = ({ stats }) => {
   let answerNameList = [];
@@ -100,16 +103,15 @@ const StarStats = ({ stats }) => {
 
   return (
     <Container>
-      <Header>
-        <TitleContainer>
-          <p>{stats.questionType}</p>
-          <h2>
-            {stats.questionNum}. {stats.questionTitle}
-          </h2>
-          <h4>{stats.questionSummary}</h4>
-        </TitleContainer>
+      <StatsHeaderContainer>
+        <Title
+          questionType={stats.questionType}
+          questionNum={stats.questionNum}
+          questionTitle={stats.questionTitle}
+          questionSummary={stats.questionSummary}
+        />
         <Average>평점: {stats.questionAvg}</Average>
-      </Header>
+      </StatsHeaderContainer>
       <ChartContainer>
         <ApexCharts
           options={options.options}
@@ -124,62 +126,34 @@ const StarStats = ({ stats }) => {
 };
 
 const Container = styled.div`
+  position: relative;
   width: 47.2rem;
   height: 41.6em;
-  background: #ffffff;
+
+  background: ${({ theme }) => theme.backgroundColor};
   box-shadow: 0px 0px 7px 3px rgba(0, 0, 0, 0.25);
   border-radius: 10px;
-  position: relative;
-`;
-
-const Header = styled.div`
-  display: flex;
-  justify-content: space-between;
-  padding: 3rem 3rem 0 3rem;
-`;
-
-const TitleContainer = styled.div`
-  p {
-    ${fonts.Body3}
-    font-weight: 400;
-    font-size: 1rem;
-    line-height: 1.2rem;
-    margin: 0;
-  }
-  h2 {
-    ${fonts.Body1}
-    font-weight: 700;
-    font-size: 1.6rem;
-    line-height: 1.9rem;
-    margin: 0;
-    margin-top: 0.6rem;
-  }
-  h4 {
-    ${fonts.Body3}
-    font-weight: 400;
-    font-size: 1.2rem;
-    line-height: 1.4rem;
-    margin: 0;
-    margin-top: 0.7rem;
-  }
 `;
 
 const Average = styled.div`
-  ${fonts.Body3}
-  height: 3.3rem;
-  width: 7.3rem;
-  font-weight: 500;
-  font-size: 1.2rem;
-  line-height: 1.4rem;
-  background: ${({ theme }) => theme.lightMainColor};
-  border-radius: 1rem;
   display: flex;
   justify-content: center;
   align-items: center;
+
+  height: 3.3rem;
+  width: 7.3rem;
+
+  ${fonts.Body3}
+  font-weight: 500;
+  font-size: 1.2rem;
+  line-height: 1.4rem;
+
+  background: ${({ theme }) => theme.lightMainColor};
+  border-radius: 1rem;
 `;
 
 const ChartContainer = styled.div`
-  margin: 1rem 0 0 1rem;
+  margin: 1rem 0 0 1.5rem;
 `;
 
 export default StarStats;
