+++ src/components/stats/statsMainList/screen/StatsMainListScreen.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import styled from "styled-components";
+
 import StatsList from "../components/StatsList";
 
 const StatsMainListScreen = () => {
